Split llvm anaylsis h and cpp files
